refactor(ts): Phase 4 - Convert API routes
- Convert src/pages/api/nwc/index.js to TypeScript
- Convert src/pages/api/links/index.js to TypeScript
- Convert src/pages/api/link/[slug].js to TypeScript
- Convert src/pages/api/claim/[slug].js to TypeScript
- Add @types/uuid dependency
- Type all request bodies and responses
